LayoutTests/imported/w3c/web-platform-tests/css/css-values/hypot-pow-sqrt-serialize-expected.txt


FAIL 'pow(1,1)' as a specified value should serialize as 'calc(1)'. assert_equals: 'pow(1,1)' and 'calc(1)' should serialize the same in specified values. expected "calc(1)" but got "pow(1, 1)"
FAIL 'scale(pow(1,1))' as a specified value should serialize as 'scale(calc(1))'. assert_equals: 'scale(pow(1,1))' and 'scale(calc(1))' should serialize the same in specified values. expected "scale(calc(1))" but got "scale(pow(1, 1))"
PASS 'pow(1,1)' as a computed value should serialize as '1'.
PASS 'scale(pow(1,1))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
FAIL 'hypot(1)' as a specified value should serialize as 'calc(1)'. assert_equals: 'hypot(1)' and 'calc(1)' should serialize the same in specified values. expected "calc(1)" but got "hypot(1)"
FAIL 'scale(hypot(1))' as a specified value should serialize as 'scale(calc(1))'. assert_equals: 'scale(hypot(1))' and 'scale(calc(1))' should serialize the same in specified values. expected "scale(calc(1))" but got "scale(hypot(1))"
PASS 'hypot(1)' as a computed value should serialize as '1'.
PASS 'scale(hypot(1))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
FAIL 'sqrt(1)' as a specified value should serialize as 'calc(1)'. assert_equals: 'sqrt(1)' and 'calc(1)' should serialize the same in specified values. expected "calc(1)" but got "sqrt(1)"
FAIL 'scale(sqrt(1))' as a specified value should serialize as 'scale(calc(1))'. assert_equals: 'scale(sqrt(1))' and 'scale(calc(1))' should serialize the same in specified values. expected "scale(calc(1))" but got "scale(sqrt(1))"
PASS 'sqrt(1)' as a computed value should serialize as '1'.
PASS 'scale(sqrt(1))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
PASS 'calc(pow(1,1) + 1)' as a specified value should serialize as 'calc(2)'.
PASS 'scale(calc(pow(1,1) + 1))' as a specified value should serialize as 'scale(calc(2))'.
FAIL 'calc(pow(1,1) + 1)' as a computed value should serialize as '2'. assert_equals: '2' should round-trip exactly in computed values. expected "2" but got "1"
PASS 'scale(calc(pow(1,1) + 1))' as a computed value should serialize as 'matrix(2, 0, 0, 2, 0, 0)'.
PASS 'calc(hypot(1) * 2)' as a specified value should serialize as 'calc(2)'.
PASS 'scale(calc(hypot(1) * 2))' as a specified value should serialize as 'scale(calc(2))'.
FAIL 'calc(hypot(1) * 2)' as a computed value should serialize as '2'. assert_equals: '2' should round-trip exactly in computed values. expected "2" but got "1"
PASS 'scale(calc(hypot(1) * 2))' as a computed value should serialize as 'matrix(2, 0, 0, 2, 0, 0)'.
PASS 'calc(sqrt(1) - 1)' as a specified value should serialize as 'calc(0)'.
PASS 'scale(calc(sqrt(1) - 1))' as a specified value should serialize as 'scale(calc(0))'.
PASS 'calc(sqrt(1) - 1)' as a computed value should serialize as '0'.
PASS 'scale(calc(sqrt(1) - 1))' as a computed value should serialize as 'matrix(0, 0, 0, 0, 0, 0)'.
